pyproject.toml
[project]
name = "function-alchemy"
version = "0.1.0"
description = "Transform any LLM into a function-calling powerhouse! 🧪"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.3.0",
    "datasets>=3.2.0",
    "peft>=0.14.0",
    "python-dotenv>=1.0.1",
    "transformers>=4.48.2",
    "trl>=0.14.0",
    "wandb>=0.19.5",
]

[tool.hatch.version]
path = "src/function_alchemy/__about__.py"

[dependency-groups]
dev = [
    "hatch>=1.14.0",
    "pytest-asyncio>=0.25.1",
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/function_alchemy"]

[tool.ruff]
lint.select = ["E", "F"]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py310"
line-length = 120

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
